requirements.txt
pandas==2.2.1
ShopifyAPI==12.4.0
